Improved email not found error message
We changed the devise partial so that the email error explanation template from the devise.en.yml file is not used. We also added our own error message in the en.yml file to make it more human, which was the main purpose of this change. Please note that further devise error message changes will likely require adding into the en.yml file and it's worth double checking the rails log to ensure this is indeed the file rails checks before the devise template ones.