Refactor layers overlay and API to improve performance and fix some edge-case issues.
The largest refactor has been in layers_overlay which is now broken into a few methods to do the rendering, and most of the state is held in the layers store. The layers are now loaded at request time rather than in state.json which speeds things up a bit.